Safeguarding against double destroy() on a few more widgets with popups.   Fixes #9966 !strict.
git-svn-id: http://svn.dojotoolkit.org/src/dijit/trunk@20312 560b804f-0ae3-0310-86f3-f6aa0a117693

+++ form/Select.js
@@ -289,7 +289,7 @@ dojo.declare("dijit.form.Select", [dijit.form._FormSelectWidget, dijit._HasDropD
 	},
 
 	uninitialize: function(preserveDom){
-		if(this.dropDown){
+		if(this.dropDown && !this.dropDown._destroyed){
 			this.dropDown.destroyRecursive(preserveDom);
 			delete this.dropDown;
 		}
